services/gateway/pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = "API gateway service"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
PyJWT = "^2.7"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
